Refactor: Renamed IStringService as StringService
Refactor: Renamed StringService as StringServiceImpl
Add: StringUtil
Refactor: Moved methods from StringService to StringUtil
