+++ src/main/java/com/medipol/h5210017/midterm/controller/StringController.java
@@ -1,7 +1,7 @@
 package com.medipol.h5210017.midterm.controller;
 
 import com.medipol.h5210017.midterm.model.LongestWord;
-import com.medipol.h5210017.midterm.service.StringService;
+import com.medipol.h5210017.midterm.service.string.StringServiceImpl;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
@@ -14,17 +14,17 @@ import java.util.Optional;
 @RequestMapping("string")
 public class StringController {
 
-    private final StringService stringService = new StringService();
+    private final StringServiceImpl stringServiceImpl = new StringServiceImpl();
 
     @GetMapping("/split")
     public List<String> split(@RequestParam("text") String text, @RequestParam(required = false) Optional<Character> separator) {
-        return stringService.split(text, separator.orElseGet(() -> ' '));
+        return stringServiceImpl.split(text, separator.orElseGet(() -> ' '));
     }
 
     @GetMapping("/longest")
     public LongestWord longest(@RequestParam("text") String text) {
-        List<String> wordList = stringService.split(text, ' ');
-        String longest = stringService.findLongestWord(wordList);
+        List<String> wordList = stringServiceImpl.split(text, ' ');
+        String longest = stringServiceImpl.findLongestWord(wordList);
         return new LongestWord(longest, longest.length());
     }
 }
